// src/type/jcg-topic-enum.ts
export enum JCG_TOPIC {
  DB_ARCHITECTURE = 'DB_ARCHITECTURE',
  DESIGN_PATTERN = 'DESIGN_PATTERN',
  SOFTWARE_DIAGRAM = 'SOFTWARE_DIAGRAM',
  INTEGRATION_TEST = 'INTEGRATION_TEST',
  MODULE_COUPLING_COHESION = 'MODULE_COUPLING_COHESION',
  SOFTWARE_DEVELOPMENT_MODEL = 'SOFTWARE_DEVELOPMENT_MODEL',
  TESTING_TYPES = 'TESTING_TYPES'
}